requirements.txt
# Core language utilities
numpy>=1.23.0
pandas>=1.5.0

# Database
SQLite

# Embedding models (HuggingFace + PyTorch backend)
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Machine learning (for clustering, PCA, etc.)
scikit-learn>=1.2.0
umap-learn>=0.5.3

# Graph analysis (citation DAG)
networkx>=3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.2
plotly>=5.13.0

# CLI
typer[all]>=0.9.0

# Config & utilities
PyYAML>=6.0
pydantic>=1.10.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Code quality (dev only)
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0

